Refactoring access methods into DAOs

+++ LAB1/dst1_0626885/src/dst1/model/ClusterDAO.java
@@ -10,7 +10,27 @@ public class ClusterDAO {
 		this.entityManager = entityManager;
 	}
 	
-	public void removeCluster(Cluster cluster_) {
+	/**
+	 * Saves a cluster to the persistence context
+	 * @param cluster the cluster to save
+	 * @return true if saved successfully, false otherwise
+	 */
+	public boolean saveCluster(Cluster cluster) {
+		entityManager.persist(cluster);
+		return true;
+	}
+	
+	/**
+	 * Removes a cluster from the persistence context
+	 * @param clusterId the Id of the cluster to remove
+	 * @return true if successfully removed, false otherwise
+	 */
+	public boolean removeCluster(Long clusterId) {
+		Cluster cluster_ = entityManager.find(Cluster.class, clusterId);
+		if(cluster_ == null)
+			return false;
+		
 		entityManager.remove(cluster_);
+		return true;
 	}
 }
